included final output manipulation, found compromise on houghcircle

+++ src/Recognizer.cpp
@@ -38,21 +38,22 @@ void Recognizer::preprocess() {
     // adaptiveThreshold(gray_blur, thresh, 255, ADAPTIVE_THRESH_GAUSSIAN_C, THRESH_BINARY_INV, 11, 1);
 
     // Otsu thresholding
-    cv::threshold(gray_blur, thresh, 0, 255, CV_THRESH_BINARY | CV_THRESH_OTSU);
+    double otsu_tresh = threshold(gray_blur, thresh, 0, 255, CV_THRESH_BINARY | CV_THRESH_OTSU);
+
+    double high_thresh_val = otsu_tresh, lower_thresh_val = otsu_tresh / 2;
+    cout << "high: " << high_thresh_val << " low: " << lower_thresh_val << endl;
 
     namedWindow("Thresh", WINDOW_NORMAL);
     imshow("Thresh", thresh);
     waitKey(0);
 
-    // Closing
-    // It is obtained by the dilation of an image followed by an erosion.
-    // dst = close( src, element ) = erode( dilate( src, element ) )
-    // Useful to remove small holes (dark regions).
-
+    // Opening
     Mat kernel = Mat::ones(3, 3, CV_8U);
 
-    Mat closing;
-    morphologyEx(thresh, closing, MORPH_CLOSE, kernel, Point(-1, -1), 10);
+    Mat opening;
+
+    morphologyEx(thresh, opening, MORPH_OPEN, kernel, Point(-1, -1), 5);
+
     /*
     src	Source image. The number of channels can be arbitrary. The depth should be one of CV_8U, CV_16U, CV_16S, CV_32F or CV_64F.
     dst	Destination image of the same size and type as source image.
@@ -63,8 +64,16 @@ void Recognizer::preprocess() {
     borderType	Pixel extrapolation method, see cv::BorderTypes
     borderValue	Border value in case of a constant border. The default value has a special meaning.
     */
-    namedWindow("Closing", WINDOW_NORMAL);
-    imshow("Closing", closing);
+    namedWindow("Opening", WINDOW_NORMAL);
+    imshow("Opening", opening);
+    waitKey(0);
+
+    // Canny
+
+    Mat edges;
+    Canny(img_gray, edges, high_thresh_val * 3, lower_thresh_val * 4, 3, true);
+    namedWindow("Canny", WINDOW_NORMAL);
+    imshow("Canny", edges);
     waitKey(0);
 
     // DEBUG
@@ -105,13 +114,19 @@ void Recognizer::preprocess() {
     vector<Vec3f> coin;
     // A vector data type to store the details of coins.
 
-    int max_radius = max(image.size().height / 3, image.size().width / 3);
+    int max_radius = min(image.size().height, image.size().width) / 4;
+    int min_dist = min(image.size().height, image.size().width) / 5.5;
+    int min_radius = min_dist * 1.2;
 
     // OK for more coins with occlusions
     //HoughCircles(closing,coin,CV_HOUGH_GRADIENT,1,50,1000,10,10,0);
 
     // OK for first image
-    HoughCircles(closing, coin, CV_HOUGH_GRADIENT, 1, 100, 1000, 20, 10, max_radius);
+    // PERFECT ON 1.jpg:
+    //HoughCircles(edges, coin, CV_HOUGH_GRADIENT, 1, 500, high_thresh_val, 10, 100, max_radius);
+    // DECENT ON everything
+    HoughCircles(edges, coin, CV_HOUGH_GRADIENT, 1.5, min_dist, high_thresh_val * 3, 30, min_radius, max_radius);
+
     // Argument 1: Input image mode
     // Argument 2: A vector that stores 3 values: x,y and r for each circle.
     // Argument 3: CV_HOUGH_GRADIENT: Detection method.
@@ -122,10 +137,15 @@ void Recognizer::preprocess() {
     // Argument 8: Minimum radius to be detected. Put zero as default
     // Argument 9: Maximum radius to be detected. Put zero as default
 
-    int l = coin.size();
-    // Get the number of coins.
+    cout << "\n The number of coins is: " << coin.size() << "\n\n";
 
-    cout << "\n The number of coins is: " << l << "\n\n";
+    // todo: remove return
+    if (coin.size() == 0)
+        return;
+
+    // DEBUG
+    for (size_t i = 0; i < coin.size(); i++)
+        cout << i << ": " << coin[i];
 
     // clean the directory used to perform classification by Python script
     system("exec rm -r images/detect/*");
@@ -133,25 +153,33 @@ void Recognizer::preprocess() {
 
     // To draw and save as images the detected circles.
 
+    cout << "Saving coins..." << endl;
     for (size_t i = 0; i < coin.size(); i++) {
+        // get center
         Point center(cvRound(coin[i][0]), cvRound(coin[i][1]));
-        // Detect center
-        // cvRound: Rounds floating point number to nearest integer.
+        // get the radius
         int radius = cvRound(coin[i][2]);
-        // To get the radius from the second argument of vector coin.
 
-        Mat single_coin = image(Rect(center.x - radius, center.y - radius, radius * 2, radius * 2));
+        cout << " Save: center location for circle " << i + 1 << " : " << center << "\n Diameter : " << 2 * radius << "\n";
+
+        // check if subimages are inside the original image (detect circle can go outside it)
+        int x1 = max(0,center.x - radius);
+        int x2 = max(0,center.y - radius);
+        int x3 = min(image.size().width, center.x + radius);
+        int x4 = min(image.size().height, center.y + radius);
+
+        // create coin subimage
+        Mat single_coin = image(Rect(Point(x1, x2), Point(x3, x4)));
+
         coins_img.push_back(single_coin);
 
-        // Save the coin frame
+        // save the coin frame
         imwrite("images/detect/coins/" + to_string(i) + ".jpg", single_coin);
 
-        namedWindow("Coin Crop", WINDOW_NORMAL);
-        imshow("Coin Crop", single_coin);
-        waitKey(0);
+        //namedWindow("Coin Crop", WINDOW_NORMAL);
+        //imshow("Coin Crop", single_coin);
+        //waitKey(0);
     }
-
-    cout << "Saving complete!" << endl;
 
     // draw circles, this must be done after, otherwise we save images with other circles drawn
     for (size_t i = 0; i < coin.size(); i++) {
@@ -166,7 +194,6 @@ void Recognizer::preprocess() {
         //  To get the circle outline.
         circle(image, center, radius, Scalar(0, 0, 255), 3, 8, 0);
         // circle outline
-        cout << " Center location for circle " << i + 1 << " : " << center << "\n Diameter : " << 2 * radius << "\n";
     }
     cout << "\n";
 
@@ -192,6 +219,11 @@ void Recognizer::preprocess() {
 
 vector<string> Recognizer::output() {
 
+    if (coins_img.size() == 0){
+        cout << "No coins detected, aborting recogniction" << endl;
+        return pred;
+    }
+
     system("python test.py");
     cout << endl;
 
@@ -201,11 +233,9 @@ vector<string> Recognizer::output() {
     if (!f.is_open()) {
         cout << "error opening file." << endl;
     }
-
     else{
         string single_pred;
         while(std::getline(f, single_pred, ',')) {
-            cout << "Found: " << single_pred << endl;
             pred.push_back(single_pred);
         }
     }
